refactor: 크루, 사용자 데이터 형식 Integer -> Long

+++ project/back/health/src/main/java/com/ssafy/health/domain/crew/dto/analysis/CrewData.java
@@ -8,8 +8,8 @@ import lombok.Data;
 @Builder
 public class CrewData {
     @JsonProperty("crew_id")
-    private int crewId;
+    private Long crewId;
     private ScoreData score;
     @JsonProperty("crew_sports")
-    private int crewSports;
+    private Long crewSports;
 }
